Fixing tests

+++ tests/setup.js
@@ -1,10 +1,18 @@
 const mongoose = require('mongoose');
 
-beforeEach('test hooks', (done) => {
+beforeEach('starts mongodb connection', (done) => {
   /* Connect to the DB before each test */
-  mongoose.connect(process.env.MONGO_URI, () => {
+  mongoose.connect(process.env.MONGO_URI, { useNewUrlParser: true }, async () => {
     /* Drop the DB */
-    mongoose.connection.db.dropDatabase();
+    if (mongoose.connection.db !== undefined) {
+      await mongoose.connection.db.dropDatabase();
+    }
+
     done();
   });
+});
+
+afterEach('closes mongodb connection', (done) => {
+  mongoose.connection.close();
+  done();
 });
